Safari-611.1.10/LayoutTests/http/tests/eventloop/queue-task-across-cross-site-frames-expected.txt

This tests the order by which tasks are scheduled across documents that are not similar origins.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS logs.join(", ") is "1, 2, 3, 4, 5"
PASS crossOriginLogs.join(", ") is "10, 11, 12"
PASS internals.hasSameEventLoopAs(frame1.contentWindow) is true
PASS internals.hasSameEventLoopAs(frame2.contentWindow) is false
PASS internals.hasSameEventLoopAs(frame3.contentWindow) is true
PASS internals.hasSameEventLoopAs(frame4.contentWindow) is false
PASS successfullyParsed is true

TEST COMPLETE
